Extracted the post item code into its own file for future use.
